+++ app/login/page.tsx
@@ -1,97 +1,8 @@
-"use client"
-import { cn } from "@/lib/utils"
-import { Button } from "@/components/ui/button"
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card"
-import { GalleryVerticalEnd, GithubIcon } from "lucide-react"
-import Link from "next/link"
-import { useGoogleSignIn } from "@/lib/auth/client"
-
-const LoginForm = () =>  {
+import LoginForm from '@/components/auth/LoginForm'
+const page = () => {
   return (
-   <div className="flex min-h-svh flex-col items-center justify-center gap-6 p-6 md:p-10">
-      <div className="flex w-full max-w-sm flex-col gap-6">
-        <Link href="/" className="flex items-center gap-2 self-center font-medium">
-          <div className="bg-primary text-primary-foreground flex size-6 items-center justify-center rounded-md">
-            <GalleryVerticalEnd className="size-4" />
-          </div>
-          Smeal.
-        </Link>
-        <div className={cn("flex flex-col gap-6")}>
-          <Card className="shadow-neutral-50">
-            <CardHeader className="text-center">
-              <CardTitle className="text-xl">Welcome to Smeal.</CardTitle>
-              <CardDescription>
-                Please sigin-in to get started with smeal.
-              </CardDescription>
-            </CardHeader>
-            <CardContent>
-              <div>
-                <div className="grid gap-6">
-                  <div className="flex flex-col gap-4">
-                    <Button variant="outline" className="w-full">
-                      <GithubIcon/>
-                      Login with Github
-                    </Button>
-                    <Button onClick={() => useGoogleSignIn} variant="outline" className="w-full">
-                      <svg width="18" height="18" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 48 48"><path fill="#FFC107" d="M43.611,20.083H42V20H24v8h11.303c-1.649,4.657-6.08,8-11.303,8c-6.627,0-12-5.373-12-12c0-6.627,5.373-12,12-12c3.059,0,5.842,1.154,7.961,3.039l5.657-5.657C34.046,6.053,29.268,4,24,4C12.955,4,4,12.955,4,24c0,11.045,8.955,20,20,20c11.045,0,20-8.955,20-20C44,22.659,43.862,21.35,43.611,20.083z"></path><path fill="#FF3D00" d="M6.306,14.691l6.571,4.819C14.655,15.108,18.961,12,24,12c3.059,0,5.842,1.154,7.961,3.039l5.657-5.657C34.046,6.053,29.268,4,24,4C16.318,4,9.656,8.337,6.306,14.691z"></path><path fill="#4CAF50" d="M24,44c5.166,0,9.86-1.977,13.409-5.192l-6.19-5.238C29.211,35.091,26.715,36,24,36c-5.202,0-9.619-3.317-11.283-7.946l-6.522,5.025C9.505,39.556,16.227,44,24,44z"></path><path fill="#1976D2" d="M43.611,20.083H42V20H24v8h11.303c-0.792,2.237-2.231,4.166-4.087,5.571c0.001-0.001,0.002-0.001,0.003-0.002l6.19,5.238C36.971,39.205,44,34,44,24C44,22.659,43.862,21.35,43.611,20.083z"></path></svg>
-                      Login with Google
-                    </Button>
-                  </div>
-                  {/* <div className="after:border-border relative text-center text-sm after:absolute after:inset-0 after:top-1/2 after:z-0 after:flex after:items-center after:border-t">
-                    <span className="bg-card text-muted-foreground relative z-10 px-2">
-                      Or continue with
-                    </span>
-                  </div>
-                  <div className="grid gap-6">
-                    <div className="grid gap-3">
-                      <Label htmlFor="email">Email</Label>
-                      <Input
-                        id="email"
-                        type="email"
-                        placeholder="m@example.com"
-                        required
-                      />
-                    </div>
-                    <div className="grid gap-3">
-                      <div className="flex items-center">
-                        <Label htmlFor="password">Password</Label>
-                        <a
-                          href="#"
-                          className="ml-auto text-sm underline-offset-4 hover:underline"
-                        >
-                          Forgot your password?
-                        </a>
-                      </div>
-                      <Input id="password" type="password" required placeholder=""/>
-                    </div>
-                    <Button type="submit" className="w-full">
-                      Login
-                    </Button>
-                  </div>
-                  <div className="text-center text-sm">
-                    Don&apos;t have an account?{" "}
-                    <a href="#" className="underline underline-offset-4">
-                      Sign up
-                    </a>
-                  </div> */}
-                </div>
-              </div>
-            </CardContent>
-          </Card>
-          <div className="text-muted-foreground *:[a]:hover:text-primary text-center text-xs text-balance *:[a]:underline *:[a]:underline-offset-4">
-            By clicking continue, you agree to our <a href="#">Terms of Service</a>{" "}
-            and <a href="#">Privacy Policy</a>.
-          </div>    
-        </div>
-      </div>
-    </div>
-
+    <LoginForm />
   )
 }
-export default LoginForm
+
+export default page
